# service-api-gateway-node/docs/openapi.yaml
openapi: 3.0.3
info:
  title: FairGiG Grievance Node API
  description: |
    Express service for grievances, analytics, and certificate endpoints.
    Interactive docs are available at /api-docs.
    Grievances, analytics, certificates, and worker community require `Authorization: Bearer` (Supabase access token).
  version: 0.2.0
servers:
  - url: http://localhost:5000
    description: Local
  - url: /
    description: Same-origin host
...
paths:
...
  /api/v1/me:
    get:
      tags: [Auth]
      summary: Current user + profile
      description: |
        Send `Authorization: Bearer <access_token>` from Supabase Auth (same token the frontend gets after sign-in).
      operationId: getMe
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User and optional profile row
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeMe"
        "401":
          description: Missing or invalid token
        "503":
          description: Server missing Supabase env
...
  /api/v1/auth/signup:
    post:
      tags: [Auth]
      summary: Register with email and password
      description: |
        Creates a Supabase Auth user. Profile rows are created by DB trigger (`worker` role only).
        If email confirmation is enabled, `session` may be null until the user confirms.
        Response includes `profile` (with `role`) when the server can read `public.profiles` (requires `SUPABASE_SERVICE_ROLE_KEY` on the API).
      operationId: postAuthSignup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
                full_name:
                  type: string
                  description: Stored in user metadata; copied to profiles by trigger when present.
                city_zone:
                  type: string
      responses:
        "201":
          description: User created (session may be null if confirmation required)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeAuthSession"
        "400":
          description: Validation or Supabase signup error
        "503":
          description: Server missing SUPABASE_ANON_KEY
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Sign in with email and password
      description: |
        Returns a Supabase session and `profile` (including `role`) for the frontend when profile lookup is available.
        Use `access_token` as `Authorization: Bearer` for `/api/v1/me`.
      operationId: postAuthLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        "200":
          description: Session issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeAuthSession"
        "400":
          description: Bad request
        "401":
          description: Invalid credentials, unconfirmed email, or no session
        "503":
          description: Server missing SUPABASE_ANON_KEY
...
  /api/grievances:
    get:
      tags: [Grievances]
      summary: List grievances
      operationId: listGrievances
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: worker_id
          description: Verifier/advocate/analyst only; ignored for workers (always scoped to caller).
          schema:
            type: string
        - in: query
          name: platform
          schema:
            type: string
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum: [open, escalated, resolved]
        - in: query
          name: tag
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        "401":
          description: Missing or invalid token
        "403":
          description: No profile
        "200":
          description: Grievances list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeGrievanceList"
    post:
      tags: [Grievances]
      summary: Create grievance
      operationId: createGrievance
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [platform, category, description]
              properties:
                worker_id:
                  type: string
                  description: Required for verifier/advocate/analyst; ignored for workers (uses JWT subject).
                platform:
                  type: string
                category:
                  type: string
                description:
                  type: string
                status:
                  type: string
                  enum: [open, escalated, resolved]
                tags:
                  oneOf:
                    - type: string
                    - type: array
                      items:
                        type: string
      responses:
        "401":
          description: Missing or invalid token
        "403":
          description: No profile or worker cannot set worker_id for others
        "201":
          description: Grievance created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeGrievance"
...
  /api/community:
    post:
      tags: [Community]
      summary: Create community post
      description: Worker only; status starts pending until an advocate or analyst approves.
      operationId: createCommunityPost
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                title:
                  type: string
                body:
                  type: string
                platform:
                  type: string
                category:
                  type: string
                tags:
                  oneOf:
                    - type: string
                    - type: array
                      items:
                        type: string
      responses:
        "201":
          description: Created
        "403":
          description: Not a worker
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase user access_token (not service_role)
  schemas:
    Grievance:
      type: object
      properties:
        id:
          type: string
        worker_id:
          type: string
        platform:
          type: string
        category:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, escalated, resolved]
        tags:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
...
    EnvelopeGrievance:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            grievance:
              $ref: "#/components/schemas/Grievance"
        error:
          nullable: true
    EnvelopeGrievanceList:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Grievance"
            pagination:
              type: object
              properties:
                limit:
                  type: integer
                offset:
                  type: integer
                count:
                  type: integer
        error:
          nullable: true
...
    EnvelopeMe:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            user:
              type: object
              properties:
                id:
                  type: string
                  format: uuid
                email:
                  type: string
                  nullable: true
                phone:
                  type: string
                  nullable: true
            profile:
              nullable: true
              type: object
              properties:
                id:
                  type: string
                  format: uuid
                full_name:
                  type: string
                  nullable: true
                role:
                  type: string
                  enum: [worker, verifier, advocate, analyst]
                city_zone:
                  type: string
                  nullable: true
                created_at:
                  type: string
                  format: date-time
            earnings_verification_summary:
              nullable: true
              description: Worker-only aggregate counts over all shift logs (from earnings table).
              type: object
              properties:
                pending:
                  type: integer
                verified:
                  type: integer
                flagged:
                  type: integer
                unverifiable:
                  type: integer
                total:
                  type: integer
                last_shift_log_at:
                  type: string
                  format: date-time
                  nullable: true
        error:
          nullable: true
    AuthSession:
      type: object
      nullable: true
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
        expires_at:
          type: integer
        token_type:
          type: string
    AuthUserBrief:
      type: object
      nullable: true
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email_confirmed_at:
          type: string
          nullable: true
          format: date-time
    EnvelopeAuthSession:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            session:
              $ref: "#/components/schemas/AuthSession"
            user:
              $ref: "#/components/schemas/AuthUserBrief"
            profile:
              nullable: true
              description: From public.profiles; null if row missing or service role not configured on API
              type: object
              properties:
                id:
                  type: string
                  format: uuid
                full_name:
                  type: string
                  nullable: true
                role:
                  type: string
                  enum: [worker, verifier, advocate, analyst]
                city_zone:
                  type: string
                  nullable: true
                created_at:
                  type: string
                  format: date-time
        error:
          nullable: true